+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated `CpiAccountSet` to use native pinocchio instruction types (#247)
 - Updated `CpiBuilder` to take in InstructionData by value and a program argument instead of ctx,
   and `MakeCpi::cpi` is now infallible (#247)
-- Non-unit return data must be wrapped in `ReturnData` (#247)
+- Return data must be NoUninit instead of `BorshSerialize` (#247)
 
 ### Removed